♻️  zb: Split handshake's common code into its own mod
