+++ zbus/src/connection/handshake/mod.rs
@@ -1,6 +1,7 @@
 mod auth_mechanism;
 mod client;
 mod command;
+mod common;
 mod cookies;
 #[cfg(feature = "p2p")]
 mod server;
@@ -9,7 +10,6 @@ use async_trait::async_trait;
 #[cfg(unix)]
 use nix::unistd::Uid;
 use std::{collections::VecDeque, fmt::Debug};
-use tracing::{instrument, trace};
 use zvariant::Str;
 
 #[cfg(windows)]
@@ -21,6 +21,7 @@ use super::socket::{BoxedSplit, ReadHalf, WriteHalf};
 pub use auth_mechanism::AuthMechanism;
 use client::ClientHandshake;
 use command::Command;
+use common::HandshakeCommon;
 use cookies::Cookie;
 pub(crate) use cookies::CookieContext;
 #[cfg(feature = "p2p")]
@@ -121,121 +122,6 @@ fn sasl_auth_id() -> Result<String> {
     };
 
     Ok(id)
-}
-
-// Common code for the client and server side of the handshake.
-#[derive(Debug)]
-pub struct HandshakeCommon {
-    socket: BoxedSplit,
-    recv_buffer: Vec<u8>,
-    cap_unix_fd: bool,
-    // the current AUTH mechanism is front, ordered by priority
-    mechanisms: VecDeque<AuthMechanism>,
-}
-
-impl HandshakeCommon {
-    /// Start a handshake on this client socket
-    pub fn new(socket: BoxedSplit, mechanisms: VecDeque<AuthMechanism>) -> Self {
-        Self {
-            socket,
-            recv_buffer: Vec::new(),
-            cap_unix_fd: false,
-            mechanisms,
-        }
-    }
-
-    #[instrument(skip(self))]
-    async fn write_command(&mut self, command: Command) -> Result<()> {
-        self.write_commands(&[command]).await
-    }
-
-    #[instrument(skip(self))]
-    async fn write_commands(&mut self, commands: &[Command]) -> Result<()> {
-        let mut send_buffer =
-            commands
-                .iter()
-                .map(Vec::<u8>::from)
-                .fold(vec![], |mut acc, mut c| {
-                    acc.append(&mut c);
-                    acc.extend_from_slice(b"\r\n");
-                    acc
-                });
-        while !send_buffer.is_empty() {
-            let written = self
-                .socket
-                .write_mut()
-                .sendmsg(
-                    &send_buffer,
-                    #[cfg(unix)]
-                    &[],
-                )
-                .await?;
-            send_buffer.drain(..written);
-        }
-        trace!("Wrote all commands");
-        Ok(())
-    }
-
-    #[instrument(skip(self))]
-    async fn read_command(&mut self) -> Result<Command> {
-        self.read_commands(1)
-            .await
-            .map(|cmds| cmds.into_iter().next().unwrap())
-    }
-
-    #[instrument(skip(self))]
-    async fn read_commands(&mut self, n_commands: usize) -> Result<Vec<Command>> {
-        let mut commands = Vec::with_capacity(n_commands);
-        let mut n_received_commands = 0;
-        'outer: loop {
-            while let Some(lf_index) = self.recv_buffer.iter().position(|b| *b == b'\n') {
-                if self.recv_buffer[lf_index - 1] != b'\r' {
-                    return Err(Error::Handshake("Invalid line ending in handshake".into()));
-                }
-
-                let line_bytes = self.recv_buffer.drain(..=lf_index);
-                let line = std::str::from_utf8(line_bytes.as_slice())
-                    .map_err(|e| Error::Handshake(e.to_string()))?;
-
-                trace!("Reading {line}");
-                commands.push(line.parse()?);
-                n_received_commands += 1;
-
-                if n_received_commands == n_commands {
-                    break 'outer;
-                }
-            }
-
-            let mut buf = vec![0; 1024];
-            let res = self.socket.read_mut().recvmsg(&mut buf).await?;
-            let read = {
-                #[cfg(unix)]
-                {
-                    let (read, fds) = res;
-                    if !fds.is_empty() {
-                        return Err(Error::Handshake("Unexpected FDs during handshake".into()));
-                    }
-                    read
-                }
-                #[cfg(not(unix))]
-                {
-                    res
-                }
-            };
-            if read == 0 {
-                return Err(Error::Handshake("Unexpected EOF during handshake".into()));
-            }
-            self.recv_buffer.extend(&buf[..read]);
-        }
-
-        Ok(commands)
-    }
-
-    fn next_mechanism(&mut self) -> Result<AuthMechanism> {
-        self.mechanisms
-            .pop_front()
-            .ok_or_else(|| Error::Handshake("Exhausted available AUTH mechanisms".into()))
-    }
 }
 
 #[cfg(feature = "p2p")]
